Next steps in not creating frogger.
The Gameboard gets created and is shown.

config.js to have most of the configuration options at a central place.
Enemies dont spawn like crazy anymore and they get removed after they
 leave the canvas.

Added lodash because I like it.
The water is a 'placeholder'-tile, to visualize the area that is not a critical path.
A little bit of code cleanup has also happened